moved profile commands to resource
