+++ src/command_modules/azure-cli-profile/HISTORY.rst
@@ -2,10 +2,6 @@
 
 Release History
 ===============
-unreleased
-++++++++++
-* `account lock`: lock commands to work specifically with subscription locks
-
 2.0.13 (2017-09-22)
 +++++++++++++++++++
 * minor fixes
