+++ src/command_modules/azure-cli-resource/HISTORY.rst
@@ -5,9 +5,9 @@ Release History
 unreleased
 +++++++++++++++++++
 * group: permit --resource-group/-g options for resource group name.
+* `account lock`: lock commands to work specifically with subscription level locks
 * `group lock`: lock commands to work specifically with group level locks
 * `resource lock`: lock command to work specifically with resource level locks
-
 
 2.0.15 (2017-09-22)
 +++++++++++++++++++
