+++ src/command_modules/azure-cli-resource/azure/cli/command_modules/resource/tests/test_locks.py
@@ -180,6 +180,35 @@ class ResourceLockTests(ScenarioTest):
                  .format(lock_name, resource_group, rsrc_name, rsrc_type))
         self._sleep_for_lock_operation()
 
+    @record_only()
+    def test_subscription_locks(self):
+        lock_name = self.create_random_name('cli-test-lock', 48)
+        lock = self.cmd('az account lock create -n {} --lock-type CanNotDelete'.format(lock_name)).get_output_in_json()
+        lock_id = lock.get('id')
+
+        locks_list = self.cmd('az account lock list --query [].name').get_output_in_json()
+        self.assertTrue(locks_list)
+        self.assertIn(lock_name, locks_list)
+
+        lock = self.cmd('az account lock show -n {}'.format(lock_name)).get_output_in_json()
+        lock_from_id = self.cmd('az account lock show --ids {}'.format(lock_id)).get_output_in_json()
+
+        self.assertEqual(lock.get('name', None), lock_name)
+        self.assertEqual(lock_from_id.get('name', None), lock_name)
+        self.assertEqual(lock.get('level', None), 'CanNotDelete')
+
+        notes = self.create_random_name('notes', 20)
+        lock = self.cmd('az account lock update -n {} --notes {} --lock-type {}'
+                        .format(lock_name, notes, 'ReadOnly')).get_output_in_json()
+        self.assertEqual(lock.get('notes', None), notes)
+        self.assertEqual(lock.get('level', None), 'ReadOnly')
+
+        lock = self.cmd('az account lock update --ids {} --lock-type {}'
+                        .format(lock_id, 'CanNotDelete')).get_output_in_json()
+        self.assertEqual(lock.get('level', None), 'CanNotDelete')
+
+        self.cmd('az account lock delete -n {}'.format(lock_name))
+
     @ResourceGroupPreparer(name_prefix='cli_test_lock_commands_with_ids')
     def test_lock_commands_with_ids(self, resource_group):
         vnet_name = self.create_random_name('cli-lock-vnet', 30)
